Updated default radio structure
Modified the default platoon radio structure in acre2_settings.sqf to
more closely match the actual default platoon structure in the framework
mission.

Now each squad's fireteam has a seperate 343 channel, and then FTLs and
their SL have a squad long range net. By default SLs and FTLs are both
given a 148 for this squad channel.

There's also now a 4th radio assignment type for backpack radios. This
really only needs using for Pilots and FACs. I've set it up to be
assigned to UAV and PP by default, as UAV is a good stand in for a FAC
unit type for now.